+++ src/Modèle/Potion.java
@@ -1,7 +1,7 @@
 package Modèle;
 
 public class Potion extends ObjetSoin{
-    public Potion(int posL,int posH){
+    public Potion(double posL,double posH){
         super(10,posL,posH,"potion");
     }
 }
